# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "description": "Uniform error body.\n\n``code`` is stable and machine-readable; ``message`` is for humans and may\nchange. Transition failures carry the states involved so the client can say\nsomething useful rather than 'something went wrong'.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "currentState": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currentstate"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "requestedState": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requestedstate"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
...
      "ExamCreate": {
        "properties": {
          "code": {
            "maxLength": 40,
            "minLength": 1,
            "title": "Code",
            "type": "string"
          },
          "config": {
            "$ref": "#/components/schemas/ExamConfig"
          },
          "course": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Course",
            "type": "string"
          },
          "department": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Department",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "durationMinutes": {
            "maximum": 480.0,
            "minimum": 5.0,
            "title": "Durationminutes",
            "type": "integer"
          },
          "instructions": {
            "items": {
              "type": "string"
            },
            "title": "Instructions",
            "type": "array"
          },
          "labId": {
            "format": "uuid",
            "title": "Labid",
            "type": "string"
          },
          "questionIds": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Questionids",
            "type": "array"
          },
          "scheduledAt": {
            "format": "date-time",
            "title": "Scheduledat",
            "type": "string"
          },
          "studentIds": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Studentids",
            "type": "array"
          },
          "title": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "code",
          "course",
          "department",
          "durationMinutes",
          "scheduledAt",
          "labId"
        ],
        "title": "ExamCreate",
        "type": "object"
      },
      "ExamDetail": {
        "description": "Faculty-facing detail. Carries the answer key - never returned to a\nSTUDENT subject.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "config": {
            "$ref": "#/components/schemas/ExamConfig"
          },
          "course": {
            "title": "Course",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "createdBy": {
            "format": "uuid",
            "title": "Createdby",
            "type": "string"
          },
          "department": {
            "title": "Department",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "durationMinutes": {
            "title": "Durationminutes",
            "type": "integer"
          },
          "endsAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Endsat"
          },
          "enrolledCount": {
            "title": "Enrolledcount",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "instructions": {
            "items": {
              "type": "string"
            },
            "title": "Instructions",
            "type": "array"
          },
          "labId": {
            "format": "uuid",
            "title": "Labid",
            "type": "string"
          },
          "labName": {
            "title": "Labname",
            "type": "string"
          },
          "questionCount": {
            "title": "Questioncount",
            "type": "integer"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/QuestionOut"
            },
            "title": "Questions",
            "type": "array"
          },
          "scheduledAt": {
            "format": "date-time",
            "title": "Scheduledat",
            "type": "string"
          },
          "startsAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Startsat"
          },
          "status": {
            "$ref": "#/components/schemas/ExamStatus"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "totalMarks": {
            "title": "Totalmarks",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "code",
          "title",
          "course",
          "department",
          "status",
          "durationMinutes",
          "scheduledAt",
          "labId",
          "labName",
          "enrolledCount",
          "questionCount",
          "totalMarks",
          "config",
          "createdBy",
          "createdAt"
        ],
        "title": "ExamDetail",
        "type": "object"
      },
...
      "ExamStatus": {
        "description": "Exam lifecycle.\n\n``READY`` exists because seating the roster is a real transition with\nobservable consequences - faculty review candidate readiness in that state,\nbefore anyone is released into the paper.\n\n``ENDING`` is the deadline sweep window: the clock has run out and the\nserver is auto-submitting stragglers, but the exam is not yet closed.\nCandidate writes are refused from the moment this state is entered.",
        "enum": [
          "DRAFT",
          "SCHEDULED",
          "READY",
          "LIVE",
          "ENDING",
          "COMPLETED",
          "CANCELLED"
        ],
        "title": "ExamStatus",
        "type": "string"
      },
...
      "ExamUpdate": {
        "description": "Accepted only while the exam is DRAFT. Every field optional.",
        "properties": {
          "config": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExamConfig"
              },
              {
                "type": "null"
              }
            ]
          },
          "course": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Course"
          },
          "department": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Department"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "durationMinutes": {
            "anyOf": [
              {
                "maximum": 480.0,
                "minimum": 5.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Durationminutes"
          },
          "instructions": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instructions"
          },
          "labId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Labid"
          },
          "questionIds": {
            "anyOf": [
              {
                "items": {
                  "format": "uuid",
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Questionids"
          },
          "scheduledAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduledat"
          },
          "studentIds": {
            "anyOf": [
              {
                "items": {
                  "format": "uuid",
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Studentids"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "ExamUpdate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Page_ExamSummary_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ExamSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[ExamSummary]",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Phase 2 contract. Handlers return static examples; the shapes, status codes and operation ids are final and generate the frontend client.",
    "summary": "Server-authoritative examination platform for physical computer labs.",
    "title": "Exam Control API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/exams": {
      "get": {
        "description": "Backs the assessments table and the dashboard.",
        "operationId": "listExams",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ExamStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "labId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Labid"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ExamSummary_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Exams",
        "tags": [
          "exams"
        ]
      },
      "post": {
        "description": "Creates the exam as a DRAFT.",
        "operationId": "createExam",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExamCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Exam",
        "tags": [
          "exams"
        ]
      }
    },
    "/api/v1/exams/{exam_id}": {
      "get": {
        "description": "Faculty-facing detail, including answer keys.\n\nNever reachable by a STUDENT subject; candidates read their paper through\n``GET /sessions/{id}`` instead.",
        "operationId": "getExam",
        "parameters": [
          {
            "in": "path",
            "name": "exam_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Exam Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Exam",
        "tags": [
          "exams"
        ]
      },
      "patch": {
        "description": "Rejected with 409 unless the exam is still DRAFT.",
        "operationId": "updateExam",
        "parameters": [
          {
            "in": "path",
            "name": "exam_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Exam Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExamUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "The exam is not in a state that permits this transition."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Exam",
        "tags": [
          "exams"
        ]
      }
    },
    "/api/v1/exams/{exam_id}/cancel": {
      "post": {
        "description": "Requires a reason, which is written to the audit trail as CRITICAL.",
        "operationId": "cancelExam",
        "parameters": [
          {
            "in": "path",
            "name": "exam_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Exam Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExamCancelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "The exam is not in a state that permits this transition."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Exam",
        "tags": [
          "exams"
        ]
      }
    },
    "/api/v1/exams/{exam_id}/end": {
      "post": {
        "description": "Closes the exam early. Enters ENDING and begins the sweep.",
        "operationId": "endExam",
        "parameters": [
          {
            "in": "path",
            "name": "exam_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Exam Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamWindow"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "The exam is not in a state that permits this transition."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "End Exam",
        "tags": [
          "exams"
        ]
      }
    },
    "/api/v1/exams/{exam_id}/monitor": {
      "get": {
        "description": "Server-computed aggregates plus rows.\n\nAlso the documented fallback when the websocket is unavailable: the same\nreducer consumes this and a live frame, so losing the socket degrades the\nrefresh rate rather than correctness.",
        "operationId": "getExamMonitor",
        "parameters": [
          {
            "in": "path",
            "name": "exam_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Exam Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonitorSnapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Exam Monitor",
        "tags": [
          "exams"
        ]
      }
    },
    "/api/v1/exams/{exam_id}/results": {
      "get": {
        "description": "Ranked results. Scores are omitted while unpublished.",
        "operationId": "getExamResults",
        "parameters": [
          {
            "in": "path",
            "name": "exam_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Exam Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultsPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Exam Results",
        "tags": [
          "exams"
        ]
      }
    },
    "/api/v1/exams/{exam_id}/results/publish": {
      "post": {
        "description": "Releases scores to candidates by setting ``published_at``.",
        "operationId": "publishExamResults",
        "parameters": [
          {
            "in": "path",
            "name": "exam_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Exam Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishResultsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultsPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Publish Exam Results",
        "tags": [
          "exams"
        ]
      }
    },
    "/api/v1/exams/{exam_id}/schedule": {
      "post": {
        "description": "DRAFT to SCHEDULED, then seats the roster to reach READY.\n\nValidates the roster against lab capacity inside the same transaction. The\nfrontend performs this check too, but only as UX - a candidate can edit the\nclient, so this is where it actually holds.",
        "operationId": "scheduleExam",
        "parameters": [
          {
            "in": "path",
            "name": "exam_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Exam Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "The exam is not in a state that permits this transition."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Schedule Exam",
        "tags": [
          "exams"
        ]
      }
    },
    "/api/v1/exams/{exam_id}/sessions": {
      "get": {
        "description": "The readiness roster before start, and the monitor table during.",
        "operationId": "listExamSessions",
        "parameters": [
          {
            "in": "path",
            "name": "exam_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Exam Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SessionRow"
                  },
                  "title": "Response Listexamsessions",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Exam Sessions",
        "tags": [
          "exams"
        ]
      }
    },
    "/api/v1/exams/{exam_id}/start": {
      "post": {
        "description": "Stamps the authoritative window and releases waiting candidates.\n\nIdempotent: a retry with the same idempotency key returns the existing\nwindow rather than shifting it, and candidates already answering are left\nuntouched.",
        "operationId": "startExam",
        "parameters": [
          {
            "in": "path",
            "name": "exam_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Exam Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExamStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamWindow"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "The exam is not in a state that permits this transition."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Exam",
        "tags": [
          "exams"
        ]
      }
    },
...
  }
}